Add files via upload
Test 2 Changes. Was not able to figure out how to get delete button to work, unsure how to properly code the section of app.delete to access the array, loop through to find the proper index, and splice it. it successfully rewrites the table, its just the removal process that fails/is skipped as no error appears in the console.